Ket hop cac ban thiet ke lan 4 và sửa lỗi load lâu, + màn hình nạp, rút,...

+++ app/src/main/java/com/duong/anyquestion/classes/History.java
@@ -6,11 +6,11 @@ public class History {
     int conversation_id;
     String title;
     String name;
-    int star;
+    float star;
     String id_user;
     String id_expert;
 
-    public History(int conversation_id, String title, String name, int star, String id_user, String id_expert) {
+    public History(int conversation_id, String title, String name, float star, String id_user, String id_expert) {
         this.conversation_id = conversation_id;
         this.title = title;
         this.name = name;
